split most of the db logic into ShareRepository
